requirements.txt
Django>=4.0.4
psycopg2>=2.9.3,<2.10
uWSGI>=2.0.20,<2.1
gunicorn>=20.1.0
jsonschema>=4.5.1,<4.6
pytest>=7.1.2
pytest-django>=4.5.2
pyrebase4>=4.5.0
python-dotenv>=0.20.0
djangorestframework
markdown
django-filter
drf-spectacular
djangorestframework-simplejwt>=5.0.0